add 'type' property to slot representation
For vat-relative slot identifiers, we use type:'export' and type:'import'
instead of positive/negative numbers. This makes room for type:'promise'
coming soon.

This changes the contents of the kernel's tables: `runQueue` (which holds
"neutral" abolute slots) and `vatImports` (which used to be named
`kernelSlots`. It also changes the slot objects that liveSlots receives.