contractHost: recognize bob-first-lies success condition
This particular test is *supposed* to throw an exception, when the contract
host notices the two players have submitted different contracts to execute.
The bootstrap.js driver now treats that as a successful completion and prints
"++ DONE" to celebrate.

The kernel still prints the serialized error about 7 times. Eventually we
need a way to distinguish between user-code exceptions and kernel/liveslot
ones, and only bother printing the latter.