kernel.loadState: use correct realm for new state vector
The loadState() function was taking the supplied state vector object and
using its components directly to populate the vat tables, however this object
comes from the outer Realm (via the Controller), so it populates the tables
with wrong-Realm objects. This wasn't a problem for the vat's normal inbound
tables, but if the vat stored any devices, those device slot objects were
wrong-realm, and then they were passed to harden() which complains.

loadState() was changed to launder the state vector through JSON
stringify/parse before use, converting the objects into the correct realm
type.

A few error messages were added to improve our chances of finding this more
quickly in the future.