css-solo: build inbound side of chain interface
This uses lotion-connect to spin up a light client and follow the chain for
new blocks. We also spin up the helper tool (currently named 'sscli') as a
subprocess and tell it to listen on its REST interface. Each time a new block
is observed, we hit the REST port to fetch our mailbox, and deliver its
contents into the swingset machine.

The outbound side is incomplete, but eventually it will hit the REST port
with a request to sign and broadcast a transaction.

We use a helper instead of raw RPC commands because the swingset module (like
any cosmos-sdk module) has a particular format for its messages, driven by
the Go implementation, and we don't currently have a way to parse or encode
that way from straight javascript. In addition it also knows how to verify
Merkle proofs (I think), and knows how to sign messages properly.