More responsive CLI (#44)
* Make cli responsive.

Use of flex was requiring fixed-width cells of the command-line UI. This
switches CLI to use grids, and adds some formatting to the command entry
widgets.

* Fix CLI display when less than a page.

The history grid needs to take up the available vertical space rather
than splitting it with the help and command entry. Also, improved
the spacing and visuals for the CLI box.