update to swingset-0.0.19, fixing perf problem
This changes the persistence system to retrieve a kernelState string at the
end of each batch of processed messages, instead of giving it a dynamic
object up front. This removes a major performance bottleneck.

In the previous version, after submitting about 2000 empty calls (on my i7
2.8GHz laptop), each one was taking about 1.1ms to execute, and that was
growing at about 5% per 100 calls. The main problem was the
unserialize/append/reserialize cycle on the vat transcripts.

With this version, even after 10k calls, each one takes less than 10us to
run. The transcript is still growing as before, but we manage it more
efficiently.

Serializing the kernel state will probably be the new bottleneck. A quick
test shows that 10k messages can result in 40MB of kernel state, which takes
about 40ms to serialize.

refs https://github.com/Agoric/SwingSet/issues/94