change run-queue handling
The run-queue now contains only 'send' and 'notify' entries. The 'send'
entries specify a target (object or promise) but we no longer figure out the
vatID ahead of time: we wait until the entry comes to the front of the queue
before finding the vat to which it should be delivered (or the promise on
which it should be queued).

The 'notify' entries do not specify the flavor of resolution, they just name
a vatID and the promiseID they should be told about. We pull the resolution
type from the promise when the notify comes to the front of the queue.
Subscribing to a promise after it's been resolved queues a new notify entry.

This prepares the way for pipelining. Each time we deliver a message to a
vat, we set the `.decider` of any result promise to point at that vat too.
Then, if a 'send' gets to the front of the queue and targets such a
result (an unresolved promise that has a decider), we check to see if the
deciding vat supports pipelining or not (this is currently hardcoded to
'false'). If yes, we deliver the message to that vat (which will also
set *it's* result's decider). If no, we queue the message within the target
promise.

To enable pipelining, we need to remove the hardcoded 'false', add an option
to kernel.addGenesisVat() to enable pipelining, and write some new tests. We
should also decide whether liveSlots should support this mode (which requires
vat-side queue handling code), or whether to only enable it for the comms
vat.