add vat option to enable pipelining
Add the vat with:

```
kernel.addVat(name, setup, { enablePipelining: true });
```

and the vat will receive `dispatch.deliver(promiseID, ..)` for messages sent
to results of previously-delivered messages.

The default is for pipelining to be disabled, so these messages will be
queued in the kernel's Promise Table (inside the result promise) until the
result gets resolved by the target vat.

The comms vat is the only likely candidate for enabling pipelining. We might
not take full advantage of pipelining until we also forward
promises (`syscall.resolve(p1, ForwardTo(p2))`) aggressively, to give the
kernel an opportunity to bypass the local vats as much as possible.