change some tests into importable modules, to test under XS
This is effectively a rebase of #133 on top of the (extensive) changes
wrought by #88. But it was achieved by just inserting the same function
wrapper around the whole file and then re-running 'npm run eslint-fix' to
adjust the indentation.

Run 'git diff -w' to see that there were no unexpected semantic changes.

closes #133
refs #130