switch to 'CapData': update comms
This probably also fixes some lurking bugs w.r.t. messages that arrive in the
comms layer for promises that it knows to have been resolved earlier. We
don't yet have tests for this case, partially because I don't know how to
provoke it from the higher-level ("ocap layer" / on-top-of-liveslots) code.