rewrite state handlng: use string-string KV store
This is the first phase of #144, to narrow the API we need from a backing
store. All state mutations go through the keepers, and the keepers deal
exclusively with an object that uses string keys and string values. We still
JSON-serialize this object on demand (at the end of the batch/block).

The next phase will replace the buildVatController/buildKernel APIs to accept
a Storage object, with get/set/getRange/deleteRange methods, and have the
JSON-serialized backing store object live outside the kernel (in the host). A
subsequent phase will introduce commitCrank/commitBlock methods on that
Storage object, then a final phase will replace the JSON backing store with a
proper database.