Link to Getting Started Documentation
Let me know if I'm missing something, but it seems like we can link to our "Getting Started" documentation at the top of this README.md. That way we don't have to rewrite the instructions. If we need to make any changes, I think we should make changes to that documentation.