[Zoe] Part 2 of Switching to Zoe Invites (AtomicSwap) (#355)
This PR makes the necessary changes to the publicSwap (renamed and reworked to `atomicSwap` where the first party who offers gets an invite to match the offer that they can send to a second party). It also makes changes to one of the contract helpers: userFlow.js, which atomicSwap uses extensively.