fix: update docs and metadata to require Node.js v12.16.1 or higher
We acquired a dependency on at least 12.11 (see #837) that we don't
understand enough to relax, and we're only testing CI on 12.16.1 (the current
LTS).

So we're declaring that we require Node 12.16.1 or higher. This updates the
top-level package.json's `engines:` property, so that `yarn` will complain
immediately if you use an old Node.

It also updates a SwingSet unit test to check for the new version, as a
backup.

closes #937
refs #837 (makes it obsolete)
closes #35