fix: improve handling of orders that are consummated immediately
in the wallet, only subscribe to notifications if the offer becomes active.
in the contract, don't save offers that are immediately consummated to the book

It used to always update the counter-offers by asking zoe for fresh
getOfferStatuses(). Now it knows which offer was consumed, so it
directly removes that offer.

A little refactoring to pull out common cases.

This fixes the second problem described in #13 that included a stack trace.