docs: improve Zoe jsDoc annotations to support contract devs
Improve jsDoc annotations in Zoe and ZoeHelpers to support contract devs.

import more types
declare zcf's type in the contracts
use OfferRecord (vs. Offer) where warranted
typedef for Invite that works.
typedef for opaque handles is {}. Typedefs are apparently aliases, so they're not distinct. typedef to Object with no properties is identical to `any`, which is less helpful.

JSDoc/Typedef doesn't seem to have a way of saying a record with an arbitrary number of typed Key-Value pairs
I describe one pair and append "|any".

If Invite is typedef Payment, then VSCode shows it as Payment everywhere. We typedef it as an Object with a getAllegedBrand, so it's a distinct type.