liveslots/test-vpid-liveslots: add failing tests for local-object
and lots of debug messages

liveslots gets more precise complaints about calling a handler that should no
longer be called

with this configuration, on top of mfig's changes to HP, I observe two tests
failing: resolution to `presence` and to `local-object`, when there is one
stall between resolving the promise and sending the message. In these cases,
the HP's unfulfilledHandler is invoked despite the resolution having been
sent.