test: demonstrate an attack by a zoe contract to be fixed in facets
The bug was introduced in "filter proposal give and want by sparseKeywords"
https://github.com/Agoric/agoric-sdk/pull/1076.
It allows a grifter contract (demonstrated here) to slip assets away
from a victim one keyword at a time. The victim is left with neither
their give or their want.

The solution in facets is to reinstate the insistance that offer
safety be checked against all keywords. We provide the feature
that #1076 was aiming for by allowing reallocations that don't impact
keywords in an offer's proposal.