fix: Fix the Zoe offer-safety bug demonstrated by the grifter attack
The fix is to insist that all offers retain offer safety across all
keywords. The desired ability to rearrange values outside of the
offer's keywords can be satisfied by explicitly allowing
rearrangements that don't impact any of the proposal's keywords.

This commit includes the bug fix but not a test that demonstrates the
vulnerability.