fix: fix Zoe bug in which offer safety can be violated (#1115)
* test: demonstrate an attack by a zoe contract to be fixed in facets

The bug was introduced in "filter proposal give and want by sparseKeywords"
https://github.com/Agoric/agoric-sdk/pull/1076.
It allows a grifter contract (demonstrated here) to slip assets away
from a victim one keyword at a time. The victim is left with neither
their give or their want.

The solution in facets is to reinstate the insistance that offer
safety be checked against all keywords. We provide the feature
that #1076 was aiming for by allowing reallocations that don't impact
keywords in an offer's proposal.

* fix: Fix the Zoe offer-safety bug demonstrated by the grifter attack

The fix is to insist that all offers retain offer safety across all
keywords. The desired ability to rearrange values outside of the
offer's keywords can be satisfied by explicitly allowing
rearrangements that don't impact any of the proposal's keywords.

This commit includes the bug fix but not a test that demonstrates the
vulnerability.

* test: correct the test to pass now that the bug is fixed.

* fix: run offer safety check over the entire new allocation, not just the sparse keywords version.

Co-authored-by: Kate Sills <kate@agoric.com>