fix: replace openDetail with quoting q
The new `q` both declassifies, as the old `openDetail` did, as well as
doing a modified `JSON.stringify`, to quote the declassified contents
as placed into the thrown error's message. The unquoted data itself is
still sent to the console for better debugging if the console supports
that.

`q` differs from `JSON.stringify` in being cycle tolerant. It will print
subtrees it has already seen as "<**seen**>", collapsing both dags and
cycles.