pass setMeter and transformMetering into vat code
* babel cannot be imported from a non-Start Compartment (it wants 'fs' and
  'path'), so we build a simple string-to-string transform function in the
  Controller (which *is* in the Start Compartment) and pass it through the
  kernel into each vat's setup() function
* likewise we pass a setMeter() inwards, which lets vats control the meter
  used for globals (like Array(length) and RegExp construction)
* add a working test-within-vat

The whole pathway from vatManager to setup() to liveSlots to buildRoot()
needs to be cleaned up, there are far too many positional args floating
around in strange orders. There is a `vatPowers` field, we should use more of
that.

test-metering.js is broken again, I changed some of the log messages.

Next steps:

* fix test-metering
* rewrite zoe's metered import to follow test-within-vat
* implement vat-at-a-time metering, test it