kernel: complain (once) if metering cannot set the global meter
To enable full within-vat metering, all of the following must happen:

* the application must call tameMetering(), generally by importing
`@agoric/install-metering-and-ses` before anything else
* the vat setup function must pass 'vatPowers' through to liveSlots, and then
from liveSlots into the function that builds the root object
* the vat must evaluate the guest code through importBundle, with a suitable
transform (built out of vatPowers.transformMetering) and a `getMeter`
endowment (which uses vatPowers.setMeter)

If the application imports `@agoric/install-ses` instead of
install-metering-and-ses, the kernel will not be able to control metering of
builtins. In this case, the guest code will be metered, but the JavaScript
builtins it calls will not, so code like `new Array(1e7).map(Object.create)`
could crash the machine without being caught by a meter.

The kernel will print a warning if metering is used but the builtins are not
instrumented. We don't currently require this, because most of our test
programs do not need it, and I don't want to change a lot of files right now.
Also, injecting metering will slow things down, even if it isn't used. In the
future, we might want to require this all the time (and an XS-based solution
should be a lot faster).