fix(agoric-cli)!: use plain require() to load deploy scripts
The 'agoric deploy' command runs ther dapp's deploy scripts, giving them
captp-style access (i.e. `E(zoe).install()`) to objects on the chain and/or
the local ag-solo. Previously these were loaded with '@agoric/evaluate',
mostly so we could apply the tildot transform to them.

Rather than switch to using `import-bundle` and defining a suitable confined
environment for the deploy scripts, we now import them unconfined, with a
plain `require()` call (which can be replaced by a dynamic import expression
when we switch fully to ESM). This allows them to use `fs` and other
Start-Compartment authorities (including unrestricted Date.now and
Math.random, if it really wanted them). We can tell devs that these are
normal JS programs (with captp/E added), rather than needing to explain the
ways in which they are not.

As a result, the `agoric` tool does *not* need to run under SES.

We do not provide the tildot transform. Deploy scripts must use `E()` rather
than tildot.