fix(transform-eventual-send)!: remove old-SES code, build step
This removes `@agoric/transform-eventual-send/src/rewriter`, which was used
to build a tildot transformation function that fit into the old-SES Realm
`transforms` option. Now that we're using new SES, it is no longer needed.
The unit test which exercised it was removed too, in favor of one that simply
tests the core string-to-string transform function.

This also removes the `yarn build` step, which used `rollup` to create a
single-file distribution bundle. We're removing these bundle steps across the
codebase.