feat(swingset)!: rewrite for new SES, remove build step, more
This commit performs an extensive rewrite of SwingSet's handling of SES,
vats, transforms, and metering.

SwingSet now requires the host application to provide a SES
environment (generally by importing `@agoric/install-ses` or
`@agoric/install-metering-and-ses` as the first line of the program). It no
longer creates an (old-)SES Realm internally. This increases safety
drastically, as we no longer have to carefully prevent primal-realm objects
from slipping across the boundary to kernel space. Now there is only one
`Object` object in the entire runtime.

Each vat gets a separate Compartment via `import-bundle`.

The `buildRootObject` function of each vat, which receives arguments like `E`
and `D` (to perform eventual-sends and invoke device nodes), also receives a
`vatPowers` object. This includes three new properties: `makeGetMeter` (to
create a Meter object that the kernel will refill between cranks),
`transformMetering` (which can be given to a Compartment constructor to
inject metering code), and `transformTildot` (which supports the use of
concise eventual-send syntax in evaluated code, e.g. a REPL).

Vats are not metered yet, but enough support code and tests have been added
to make the process easier in the future. If the kernel finds "metered
globals" support (which happens in applications that import
`@agoric/install-metering-and-ses` when they begin), the kernel will use it
to improve metering protection for vat code that uses it.

The "bundle kernel source" step has been removed. `buildVatController` uses
`bundle-source` (which uses `rollup`) at runtime to turn the kernel source
graph into something that can be evaluated in a new Compartment, and then
`import-bundle` to do the evaluation.