fix(cosmic-swingset)!: use install-ses, new repl tildot
cosmic-swingset contains our two primary applications: `ag-cosmos-chain` and
`ag-solo`. These have both been updated to use `install-metering-and-ses` to
provide a globally-metered SES environment for SwingSet.

The REPL has been updated to use the Compartment API, as well as the new
transformTildot function.

This adds node-lmdb to package.json, to support the LMDB-vs-SES workaround
that requires us to import LMDB before lockdown. When that gets fixed and we
remove the import, we should remove the package.json dependency too