feat(swingset): activate metering of dynamic vats
vatManager now does a check, just after the vat finishes with a crank, to see
if a meter was active for that vat (`meterRecord`). If so, it checks the
meter for exhaustion and uses `notifyTermination` to tell the admin facet
about it. Otherwise it refills the meter. Vats get unlimited cranks, but each
crank has a limited meter budget.

The dynamic-vat creation code makes a new meterRecord, applies the
metering-transform (and endowment to let the transformed code do `getMeter`),
and prepares a `notifyTermination` callback which will queue a message to the
vatAdminWrapper with the details.

The vatAdminWrapper now manages an additional `done` promise for each dynamic
vat it manages. When it gets the `notifyTermination` message, it resolves
this promise. Whichever vat object holds the `adminNode` can retrieve this
promise and get notified when the vat dies.

The unit test exercises a vat overrunning the "allocate" meter (which is easy
to trigger without a long slow loop), and makes sure the vat doesn't respond
to future messages after it's been exhausted. The existing dynamic-vat
test (vat-admin/test-innerVat.js) was updated to install global metering,
because now all dynamic vats are metered, whether the test is exercising
metering or not, and swingset emits a warning unless global metering is
installed.

We don't currently attempt to clean up the vat in any way. We rely upon the
vat's one meter remaining exhausted and never being refilled. As long as that
remains the case, all vat code will throw (the same exception) immediately
upon any message delivery, so while the vat's liveslots code gets to run (and
result promises are rejected appropriately), the vat code itself will never
get control again. The `notifyTermination` callback will be fired each time,
but it has an internal flag to ignore the later ones.

We also don't yet have a way to preemptively terminate a vat. We could add
this pretty easily by changing `doProcess` to check a new "null or Error"
flag just before dispatching into the vat.

One missing feature is that any Promises the late vat was controlling will
remain forever unresolved. Ideally all those promises should be rejected as
soon as the vat is known to be terminated. This is the highest-priority
followup task.

A further-out task is to delete the vat from memory, decref its c-list
entries, and somehow propagate disconnect messages back to holders of
now-dangling object references.

Another remaining task is to handle state cleanup and "rewind the
transaction" better. In the present code, when a vat dies mid-way through a
crank, any syscalls it made before the meter exhausted will still get
through. External observers will see a prefix of the messages the vat would
have sent if it hadn't run out of metering budget. Ideally the entire crank
would atomically happen or not happen, which will require us to buffer those
syscalls until the crank finishes with time still on the clock. This is
visible now, but isn't too serious yet. It will become more important when we
consider allowing vats to re-start the message that killed them (using a
bigger meter).