fix: add tests & fix for ZCF handling of crashes in contract code
When contracts exceed metering limits, or throw exceptions, the
contract may fail, but exit safety should kick in, and other contracts
and Zoe should continue to function.

rejectOffer() had to stop calling complete() before asert.fail(),
because it was causing the offer to no longer be active when zoe tries
to complete() all outstanding offers. The error message from the
reject was also being lost. test-zoeHelpers had to be updated not to
expect those rejected offers to have been completed.