fix(swingset)!: change kernel API for adding vats/devices
This emphasizes passing bundles into the kernel, rather than a pre-evaluated
`setup` function. The controller is responsible for `bundleSource`, leaving
the kernel responsible for `importBundle`. This will make it easier to let a
VatManager do that importBundle in a different process (e.g. an XS
VatWorker).

`kernel.addGenesisVatSetup` was added to let unit tests do their thing with
raw `setup` functions.