fix(swingset)!: add comms/timer/vattp vats automatically
The comms vat is used to route messages from one swingset to a remote
swingset: its job is to figure out which remote machine each message should
go to, and to translate our object/promise references to ones that the remote
machine will recognize.

It uses the vattp vat to actually send the messages. Some other
application-provided vat must talk to vattp and give it a way to send and
receive messages (generally through some device).

The timer wrapper vat is used to talk to the timer device, and provide a
pleasant API.

Previously, the application was responsible for adding all three to the
swingset config table. Swingset exported functions like `getVatTPSourcePath`
to let the application discover what pathname should be passed back in to
swingset's configuration. This was a drag.

Now, these three vats are added automatically to all swingsets. Applications
no longer need to ask for them; the bootstrap function can simply reference
`vats.comms`, `vats.vattp`, and `vats.timer`.

Applications are still responsible for wiring comms and vattp together, as
well as wiring `vats.timer` to `devices.timer`. This will be changed in the
future.

cosmic-swingset, the primary customer (and the only one which uses these vats
for real) was modified to match.