feat(swingset): Add Node.js Worker (thread) -based VatManager
This adds a per-vat option to run the vat code in a separate thread, sharing
the process with the main (kernel) thread, sending VatDelivery and VatSyscall
objects over the postMessage channel. This isn't particularly useful by
itself, but it establishes the protocol for running vats in a
separate *process*, possibly written in a different language or using a
different JS engine (like XS).

This 'nodeWorker' managertype has several limitations. The shallow ones are:

* vatPowers is missing transformTildot, which shouldn't be hard to add
* vatPowers.testLog is missing, only used for unit tests so we can probably
live without it
* vatPowers is missing makeGetMeter/transformMetering (and will probably
never get them, since they're only used for within-vat metering and we're
trying to get rid of that)
* metering is not implemented at all
* delivery transcripts (and replaying one) is not yet implemented

Metering shouldn't be too hard to add, although we'll probably make it an
option, to avoid paying the instrumented-globals penalty when we aren't using
it. We also need to add proper control over vat termination (via meter
exhaustion or manually).

The deeper limitation is that nodeWorkers cannot block to wait for a
syscall (like `callNow`), so they cannot invoke devices

Remaining design issues:

* I'm not yet sure how the new manager should be organized, or where it
should be located.
* the 'local' vatManager can probably be refactored to share more code with
the new 'nodeWorker' type

The controller/kernel acquired a new `shutdown()` method, to use at the end
of unit tests that create `nodeWorker` -style workers. Without `shutdown()`,
the worker thread would keep the process from ever exiting.

The process of creating vatManagers was refactored into a single
vatManagerFactory(vatID, options). Some options were added, and others fixed:
managerType, enableSetup, enableInternalMetering.

refs #1299