test(swingset): run message-patterns serially
AVA appears to use unlimited parallelism within test files (the `-c` option
seems to only limit the number of files run in parallel, not test functions).
The swingset `test-message-patterns.js` file contains 62 tests, all of which
spin up a swingset, exchange some messages, then shut down again. When run on
their own each takes 2 or 3 seconds. But with 62 in parallel, the CPU gets
swamped and it takes multiple minutes for any one test to reach an assertion,
which triggers AVA's timeouts.

I couldn't find a way to limit within-file parallelism, so this changes the
message-pattern test to run serially (it changes `test(..)` with
`test.serial(..)`).

A future improvement is to change these tests so they share code (I think
it's the `rollup` step that's most CPU-intensive, and if we can share the
bundles between tests, we might save most of the work) and can be run in
parallel once again.