test: speed up test-message-patterns.js with buildKernelBundles
This changes test-message-patterns.js to build all vat bundles ahead of time,
once, in a `test.before()` clause, and shares both the bundles and the config
objects among all tests. This speeds up the overall test from 230s to 12s on
my machine.

It also removes lingering support for enabling/disabling pipelining, which I
needed when first implementing pipelining, but is now just in the way. And a
few noisy console.logs are gone.

We can achieve further speedups by amortizing bundling of the loopbox device
code, once the swingset config can accept device bundles in addition to
sourcePaths.