feat: accept 'description' in vat creation options, pass to slog
This puts the power/responsibility for describing the vat into the hands of
the creator. For static vats, this just says `static name=NAME` (but we
could/should allow that to be set from the config file at some point). For
dynamic vats, it's up to the caller to provide a description.

In both cases, the description is only passed to the slog.

We add a copy of the source bundle to the slog too, so it can recreate the
correct initial vat image.