fix: refactor liveSlots so it could provide vat globals
This splits up `makeLiveslots` so that the `buildRootObject` function is
provided later, after it returns. This makes it possible to delay the import
of the vat code until after liveslots has provided globals to put into its
Compartment.

This does not yet provide any globals (the `vatGlobals` which `makeLiveslots`
returns is empty). Nor does it change the vat managers to use the
`vatGlobals` in any way. But these changes should be a lot easier now.

refs #1867