fix: add WeakRef tracking to liveslots
Liveslots maintains an import count: each time a vref is mentioned in an
inbound message, liveslots increments the counter. When user-level code
forgets about the Presence object, eventually (in some future turn, probably
not during one of this vat's cranks) the finalizer callback will run, and
liveslots will send a `vatDecRef()` message to the kernel with the old count.

This will eventually be wired up with kernel-side code that maintains a
similar counter, which will be decremented, and when it reaches zero, we'll
remove the c-list entry.

The handling of Promises needs more thought. Currently we retain a strong
reference to imported Promises until they are retired (which happens in
certain cases of resolution).

refs #1872