fix: tolerate decrefs that arrive after their vat has be deleted
Finalizer callbacks run at strange times, and can occur after the vat which
dropped the import has been terminated and cleaned up (so there are no c-list
entries left to remove). Ignore these.

The unit test is necessarily probablistic. This particular sequence appears
to exercise the previously-failing path about 50% of the time. The other 50%
of the time it does not, and the test passes without accomplishing anything
useful.