fix: exercise callNow in local-worker case
This is the only type of worker which can currently handle syscall.callNow.

The `node-subprocess` case theoretically could (the child subprocess could do a
blocking read of the kernel-to-worker pipe while waiting for the result), but
the code on either side of that pipe does not yet support that mode.

The `nodeWorker` (Node.js threads) case cannot, because threads cannot block,
at least not without heroics involving `Atomic` locks around a
results-bearing `SharedArrayBuffer`.

The old `xs-worker` case could not, for the same shallow implementation
reasons as `node-subprocess`, but the new `xsnap` -based XS worker should be
capable.

closes #1617