fix: reject symbols and unenumerated properties as intended
The logic in marshal's isPassByCopyrecord and assertCanBeRemotable which was
meant to look for Symbol-named properties was bypassed by the fact that
Object.entries ignores Symbols.

This should fix that logic, and updates the tests to assert that all three
kinds of bad properties are rejected: enumerated Symbol-named, unenumerated
string-named, and unenumerated Symbol-named.

SwingSet's liveslots.js was updated to use passStyleOf instead of
mustPassByPresence, which handles the case where a vat's root object has been
marked as Remotable without an error.