fix(xs-worker): restore xs vat manager test to working order
In 94e007867 on Feb 7, the 'xs vat manager' grew a `test.skip`. (It's
not entirely clear to me why. I can't reproduce the failure as of that
version.) So in 43344ad Feb 10, we didn't notice that xs's deliver now
returns `Promise<CrankResults>` while kernel.js still expects deliver
to return `Promise<Tagged>`.

Taking out the `test.skip` results in...

    TypeError#2: deliveryResult is not iterable

at kernel.js:374 `const [status, problem] = deliveryResult;`

So this changes xs's deliver to just return the .reply part of the
CrankResults. I trust it's more or less by design that the actual
metering part is not being consumed by anything at this point.