fix: hush "replaying transcripts" message during swingset startup (#2394)
This removes one of the two noisy log messages from kernel startup. This
mostly gets noticed in tests like test-message-patterns.js which launches
dozens of swingsets in a single test. We included it because sometimes errors
occur (and get logged to the console) during transcript replay, and it wasn't
obvious when/why the vat was being invoked so early. But to be properly
useful for that, we'd need to announce the replay phase of each vat
separately, as well as the *end* of replay for each vat, multiplying the
noise by at least an order of magnitude.

Instead, just remove the message, and if/when someone needs to figure this
out, go ahead and temporarily add it back. Besides, transcript replay is
going to change drastically with the introduction of the
vat-manager-manger (#2277), which would have removed the place where this
message was emitted anyways.