fix: remove crankNumber from transcript entries
The VatManager is responsible for maintaining a transcript of all deliveries
to their vat worker, so the vat can be reloaded later. These transcript
entries have been including the `crankNumber`: a global counter indicating
how many cranks have been delivered (to all vats, not just the one for which
the transcript entry is being created).

This removes that crankNumber from the transcript:

* each vat should be independent, this global crankNumber is revealing
information about what happens in other vats
* #2400 is changing the type of `crankNumber` to a BigInt, which cannot be
serialized by the simple `JSON.stringify` used in vatKeeper.js

It also removes the now-unnecessary `kernelKeeper` argument from
`makeTranscriptManager`, and changes one test that happened to depend upon
the presence of the crankNumber field.

closes #2428