fix(marshal): add Data marker, tolerate its presence
Test that it can be added to empty and non-empty data, and that it doesn't
cause changes to the serialization of non-empty objects. We do not yet change
the handling of empty objects.

Data precludes Remotable: anything marked as Data cannot also be treated as
Remotable. This is mostly to guard against confusion inside passStyleOf.