chore: use Nat 4.0.0 in swingset (#2420)
In the mailbox device, we use a plain Number, not BigInt, for external
mailbox seqnums. Host applications call `exportMailbox()`, and are then
responsible for conveying that data to a peer application. We can make their
lives more convenient by returning something that can be serialized with a
plain JSON.stringify. We make `exportMailbox()` emit plain Numbers for the
sequence numbers, even though we store them internally as BigInts (validated
with `Nat`). Likewise we accept Numbers from the host but convert them (with
Nat) into BigInts for local processing.