fix(zoe): annotate empty objects with Data or Far as appropriate
Some unit tests used to apply `t.deepEqual(facet, {})` to test that an object
was a "marker" (a zero-method pass-by-reference object) or a Presence, rather
than an empty (data) object. We now create such markers with `Far('interface
name', {})` instead of `harden({})`, which means it's no longer deep-equal to
an empty object (Far adds a distinct prototype). I added a helper method to
assert the properties we really care about: the object is pass-by-reference,
and it has no methods.

This pulls in a subset of the changes from #2545 , which seemed easier than
iterating on that branch or applying more fixes on top. Thanks @katelynsills
for the cleanup!

At this point, zoe tests pass even when marshal.js is configured to throw
errors when it sees unmarked empty objects, which implies that we've marked
all or most of them.

refs #2018