fix(spawner): add enough Far/Data to pass tests
I'm pretty sure this is incomplete: coveredCall.js and escrow.js lack Far()
calls around clearly pass-by-reference objects. It's not obvious how to add
them, because the functions therein are turned into source strings with plain
toString (backticks), so any reference to Far will result in an
undefined-global TypeError when the string is evaluated. To include Far
properly, these need to be sent through bundleSource() instead of a simple
toString.

But this doesn't seem to cause the unit tests to break, probably because 10
of the 14 tests are marked as test.skip. I'm going to PR this anyways,
because I think spawner is vestigal and is likely to be deleted entirely
before those tests are ever re-enabled.

refs #2018