fix(swingset): let command device pass BigInt args/responses/broadcasts
Previously the "command device" used JSON.stringify/parse to pass data across
the kernel/host boundary. This could not handle `undefined`, BigInts, and
other interesting pass-by-copy datatypes.

With this change, we use `marshal` stringify/parse instead of plain JSON,
which can handle a wider variety of JavaScript datatypes. This is nearly a
superset of JSON, with the unfortunate exception of empty objects, which will
not be handled quite right until #2018 is finished. Such objects will arrive
as empty Presences (with an unusual prototype and an unenumerable
Symbol-named property), or will be rejected during serialization (because
they lack a Data or Far marker), or will cause a warning but arrive properly,
depending upon what phase of the #2018 transition we're in at the time.

The one visible limitation of using `marshal` stringify/parse is that the
data you feed to it must be `harden`ed: the command device will throw an
exception if given unhardened data. The previous JSON approach did not
require this.

refs #2559