refactor(bundle-source): Isolate transform
This isolates a transformSource function so that it can be applied for
the Endo Base-64 Zip bundle format in a subsequent change.
The change reveals that the transformSource function is necessarilly
async--it uses the Mozilla SourceMapConsumer, which in turn relies on a
pool of workers for performance.
This in turn suggests the next phase of work will be adding support for
asynchronous transforms to the Endo archiver in the Compartment Mapper.

Refactoring revealed an opportunity to parallelize transforms, which
might better leverage the source map worker pool, and also revealed a
resource leak since we did not previously release source map workers.