fix: remove incorrect assertion in multipoolAutoSwap priceAuthority
fixes #2831

priceAggregator shows that when priceQuery() returns a falsy value, createQuote()
is expected to return undefined. This happens when the comparefunction returns
false, so the promise isn't resolved.

I added a test of the priceAuthority in multipoolAutoSwap. The test fails before
the change (The first trigger gets false on the comparison)  and passes afterward.