chore(swingset): update/delete some docs
delivery.md is where we describe the vat-kernel architecture the best, so I
added the newer syscalls and fixed a few outdated parts. I didn't have time
to update the Comms Vat section thoroughly.. I added a description of the
kernel-facing clist, but the examples still pretend that the comms "local"
namespace and the kernel-facing namespace are the same. Likewise I added a
note about `syscall.resolve()` handling batches now, but did not update the
details.

design.md was really really old and inaccurate/irrelevant, so I deleted it.