fix(swingset): when a static vat dies, tolerate lack of next-of-kin
`vatTerminated()` was choking during test-syscall-failure.js, because it want
to find a Promise to resolve/reject, and static vats don't have a creator
waiting to be notified about their death.

This replaces the ugly internal error with a simple warning. The death of a
static vat means something has gone seriously wrong, and we should consider
halting the kernel, but this internal error was not the right way to display
it.