fix(swingset): schedule vat creation on the run-queue
This handles properly async vat creation, such as the process used by
subprocess-xsnap, in which the vat isn't ready for use right away.
Previously, we queued the "vat ready" message (to vatAdmin) at a random time,
which was both non-deterministic, and it left the run-queue empty, making
tests and certain runtime configurations think the kernel was idle, exiting
too early.

Now, when the vatAdminVat uses its device node to create a dynamic vat, we
merely queue a new `create-vat` event on the run-queue. Later, when this
arrives at the front of the queue, we perform a crank which starts the vat
manager and waits for it to complete. Any state changes that result from this
process (such as queueing the vat-is-ready message) get committed as a part
of this crank.

closes #2991