fix(spawner): rewrite to use createVat, strip down to bare minimum API
This removes all zoe-like functionality from the spawner, leaving a bare
install() and spawn() API. This is now effectively a wrapper around
createVat(), except that it accepts source code which exports a default
function, rather than an export named `buildRootObject`, and it does not
return the control facet (`adminNode`) to the caller.

Our two cosmic-swingset applications (chain and ag-solo) are updated to
create the spawner correctly. Dapps can continue to use the existing API
without changes.

closes #1343