chore(swingset): add transcript extraction/replay tools
The extraction tools take a kernelDB directory, or a "complete" slogfile,
and creates a .sst transcript file for a single selected vat.

(A "complete" slogfile is one that was recorded from the beginning of kernel
startup. If instead the slogfile was recorded from an application restart, it
won't include results from replayed syscalls, so it won't be enough to
reconstruct the full transcript)

Note that slogfiles did not record vatParameters until recently, so
transcripts created with extract-transcript-from-slogfile.js from older
slogfiles may lack important information.

This also adds a replay tool to build a new vat from this transcript. See
how-to-replay.md for details.

closes #3110