feat(SwingSet): change virtualObjectManager API to reduce authority
Rather than giving the VOM full control over liveslot's valToSlot/slotToVal
tables, just give it functions to do the things it needs to do:
`getSlotForVal`, `getValForSlot`, and registerEntry`. This decouples the fact
that slotToVal actually holds WeakRefs. It also makes it easier for
`makeFakeVirtualObjectManager` to return debug functions that hide this fact,
so unit tests that want to manipulate or examine the internal tables can do
so without being aware of the WeakRefs involved.

This prepares for a change to `makeFakeVirtualObjectManager` to accept a `{
weak: true }` option, which will make it use the same WeakRef pattern as
liveslots does. We'll need this to test the VOM retaining Remotables, since a
strong `slotToVal` reference would cause a false positive.