chore(swingset): add design docs for garbage collection
Adds docs/garbage-collection.md with a detailed description of the design.

There's still a lot of material to add, but this should capture a lot of the
terminology and concepts we've introduced to implement GC, plus most of the
algorithms involved. Required reading to understand how GC works in SwingSet.

refs #3106