chore(swingset): add GC Actions
This is a relatively independent portion of the kernel-side GC work. "GC
Actions" are generated by the upcoming `processRefcounts()` function that the
kernel will run after each delivery is complete. For example, if the kernel
observes an object's "reachable" refcount drop to 0 during a crank, and the
count is still zero when `processRefcounts()` runs, it will push a
`dropExports` action (for the exporting vat and dropped kref) into the GC
Actions set.

The GC Actions set is stored durably in the kernelDB, and consulted by a call
to `processNextGCAction()` just before pulling anything else off the
run-queue. This call must select the highest-priority action to take, and
check to see if the kref in question has been re-exported since the action
was enqueued (removing it from the queue if so). It then returns the action
to take, or `undefined` if there are none.

This commit adds the implementation of `processNextGCAction()`, and the unit
test which exercises its various possibilities.

It also adds a new key to the kernelDB (named `gcActions`) which holds the
durable set, and several new kernelKeeper methods to manipulate it, including
`get`, a simple `add` function, and a `get` to replace the set
wholesale (since `processRefcounts` can be more efficient if it updates the
durable set in a single operation rather than a long sequence of adds).

It does not change the behavior of the kernel: nothing yet adds GC
Actions (this will happen in `processRefcounts()`, nor does anything example
the durable set to act upon any queued actions (this will happen in
`getNextMessage()`).

refs #3109