chore(swingset): add gc-actions.js and related tests
This is a relatively independent portion of the kernel-side GC work. "GC
Actions" are generated by the upcoming `processRefcounts()` function that the
kernel will run after each delivery is complete. For example, if the kernel
observes an object's "reachable" refcount drop to 0 during a crank, and the
count is still zero when `processRefcounts()` runs, it will push a
`dropExports` action (for the exporting vat and dropped kref) into the GC
Actions set.

The GC Actions set is stored durably in the kernelDB, and consulted by a call
to `processNextGCAction()` just before pulling anything else off the
run-queue. This call must select the highest-priority action to take, and
check to see if the kref in question has been re-exported since the action
was enqueued (removing it from the queue if so). It then returns the action
to take, or `undefined` if there are none.

This commit adds the implementation of `processNextGCAction()`, and the unit
test which exercises its various possibilities. It does not touch the kernel:
a later commit will implement the durable set, the `processRefcounts()` code
that populates it, and the run-queue code that uses `processNextGCAction` to
extract actions from it.

refs #2946