chore(swingset): make addExport() increment the refcount
`kernel.addExport()` is used to pretend that a vat has exported some object,
so we can get a kref on it and e.g. submit exomessages with
`kernel.queueToKref`. If a real vat had imported this kref, that vat's c-list
would maintain a reference to it, keeping the refcount non-zero and
inhibiting GC. But if it's merely a unit test that remembers the kref, there
is no importing c-list to provide the refcount, and the object is in danger
of being collected before the test can finish interacting with it.

This commit changes `kernel.addExport()` to increment the refcount by one,
replacing the missing importing vat.

It also increments the refcount on slots of `kpResolution`-retrieved promise
resolution data, for the same reason.