chore(swingset): refactor addKernelObject, add deleteKernelObject
The kernelKeeper `addKernelObject` method was updated to accept an `id=`
override, to simplify some upcoming unit tests. `deleteKernelObject` was
added, which isn't called yet but the upcoming GC changes will invoke it when
GC allows a kernel object to be deleted. It's also a placeholder for #2069
auxdata to be deleted.