fix(swingset): fix two tests which failed under XS GC
test-controller.js uses `c.step()` to examine the kernel one crank at a time,
but was written before some of those cranks are really GC actions. XS seems
to do a pretty thorough job of GC when asked, so the kernel has several to
do, and the test did not see the regular vat delivery progress it was
expecting. The fix is to step through all GC actions before taking the one
vat-delivery `c.step()` it intended to do.

test-promises.js "refcount while queued" failed because the bootstrap method
under test did not pin all vats in place. As a result, by the time the test
tried to `c.queueToVatExport()` a message to vat-right, vat-right's root
object had been collected. The fix is easy: make `bootstrap()` pin
`vats.right` in place.

The interesting question is why this did not fail under Node.js
(`managerType='local'`). Apparently Node.js is less inclined to release
objects, even when a full `gc()` is invoked. I suspect some pattern of
definition hoisting or optimization is keeping the `vats` variable around
longer than the source code would suggest. Might be related to #3240, but I
haven't been able to find a `gc()` mode that causes Node.js to collect it.