feat(swingset): comms: track lastSent seqnum for each object
This adds a DB key for each (remoteID, object) pair, which remembers the
outbound sequence number of the message which exported or re-exported the
object. Each time an outbound message to a downstream peer could grant access
to the object, we must update this counter.

This gives the upcoming GC handler function (the one which receives
`dropExport` from the downstream peer) enough information to figure out
whether their drop was "informed" (sent *after* the most recent export), or
"uninformed" (sent *before*), so it can ignore uninformed drops. This will
happen when the outbound (re-)exporting `send()` crosses the inbound `drop`
on the wire.