refactor(swingset): clean up comms dispatch, log errors
Errors during comms processing were not being logged very usefully, and
are more likely to indicate a bug in the comms code than to be something
provoked by userspace, so log any errors during `dispach()` before
re-throwing them.

Remove a number of helper functions in preparation for changes to the way GC
deliveries are handled.