fix(swingset): comms: remove deleteToRemoteMapping
When the comms vat resolves a promise, it begins a two-part retirement
process. Right after it sends the `resolve` message to the remote, it deletes
the outbound half of the c-list, because this promise ID is partially
retired, and the comms vat won't be referencing it in any outbound messages
again.

The inbound half is retained until the other side acknowledges the resolution
message (i.e. some inbound messages arrives which demonstrates awareness of
the `notify`, by including an `ackSeqNum` at least as large as the `seqNum`
of the outbound `notify`).

This changes the code to retain both sides of the c-list entry until the ack
is received. I think this winds up being slightly cleaner, and most
importantly it retains our ability to map local-ref to remote-ref, which
enables an upcoming change to `deleteRemoteMapping` / `deleteKernelMapping`
to only take a single argument (lref, not lref+rret), which is a lot cleaner,
and avoids some error cases.

refs #3306