feat: ContractGovernor manages parameter updating for a contract
ContractGovernor is an ElectionManager that knows its Registrar, and
starts up and manages one contract. The managed contract returns a
facet that describes the managed parameters to the governor. The
ContractGovernor returns a facet that allows the creation of new Ballot
questions to ask the electorate to approve changing the values of
those parameters.

There's a README.md that gives on overview of the interdependent
APIs.

The swingsetTest exercise the ContractGovernor over a trivial
governedContract that shows how a contract can configure its
parameters, and the methods that have to be available in order for
everything to be found and verified. The tests include a verification
step that cross-checks the Registrar, ElectionManager (i.e.
ContractGovernor in this case), and Ballots to see that they are
consistent.

The API of Ballot and others is changed to bundle the details of a
question into BallotSpec. This affects some of the unit tests.