fix(agoric-cli): use 'yarn workspaces' instead of hard-coded list
When `agoric install` is deciding which dapp subdirectories to examine, call
`yarn workspaces info` to the the real list from the dapp's package.json,
instead of probing for a hardcoded set. And don't require those
subdirectories to have a yarn.lock file.

If the dapp's package.json doesn't indicate the use of workspaces, just use
whatever subset of the original hard-coded subdirectories (contract/, api/,
ui/ etc) that have readable package.jsons.

Also use `yarn workspaces info` when examining agoric-sdk to see what
packages need to be targets of symlinks, instead of grabbing every subdir of
`packages/` and `golang/` (some of which may be spurious).

This helps testnet-load-generator (which uses `loadgen/`, not `api/`) to get
the right versions of all packages when it creates bundles.