fix(vats): vat-ibc must use LegacyMap, not Store, to hold a Set
The recent change to make Stores accept only "Passable" objects means that
you can't put a Set in one anymore. The IBC code was holding Sets of Promises
in a Store, and that now fails. This caused the chain to halt at startup,
during initialization and bootstrap.

The fix is to use a "LegacyMap" instead of a "Store".

fixes #3621