fix(swingset): rename snapshot-related DB keys to be "local"
We're defining the `local.*` namespace as the portion of the DB that is
excluded from consensus (the #3442 kernel activity hash). Vat snapshots are a
local concern: each member of a consensus machine can decide for itself
if/when to make a snapshot. So both the vat->snapshot and the snapshot->vats
tables must be moved to `local.`.