fix(swingset): move "kernelStats" into local/non-hashed DB space
The kernelstats are merely incidentally deterministic: unlike c-lists and
run-queues and object/promise state, stats aren't the primary thing we want
to compare across replicas in a consensus machine. And we might want to add
more in the future without causing a compatibility break. So this changes the
DB key from `kernelStats` to `local.kernelStats`, which will exclude them
from the #3442 kernel activity hash.