fix(swingset): make test-controller less sensitive to source changes
The kernel "activity hash" is sensitive to everything written into the
kvstore, including the bundled sources of built-in vats (like comms and
vatAdmin), which include upstream libraries like marshal and assert. This
makes the hash sensitive to a significant amount of code, and test-controller
was asserting a specific value. This required updating the test's
expectations on every code changes.

Instead, we now merely check that `controller.getActivityhash()` returns a
correctly-shaped string, instead of examining the actual value.

closes #3718