fix(cosmos): ensure simulated transactions can't trigger SwingSet
The former pattern of `vm.IsSimulation` was too brittle to handle
our migration to the GRPC-based transaction handlers.  Instead,
intercept the actual invocation of CallToController and prevent
it from running anything in SwingSet if in simulation mode.

Also, make sure our InfiniteGasMeter is only assigned once, so
that if we ever do Cosmos-level metering of SwingSet downcalls,
there is only one place to adjust.