feat!: add governed electorate to contractGovernor
This commit contains only the changes in the governance package. Changes to
Treasury and AMM are in separate commits, but tests won't pass without all
layers

Params in terms switch from arrays with embedded names to named records.
The electorate is passed to contracts in an invitation rather than a facet.

Dropped some tests in test-paramGovernance as the use of fakes was unsustainable.
   The two that were dropped were a happy path, and an inaccessible error path.

Added a test to show that changing the electorate works