Mutable electorate (#4168)
* feat!: add governed electorate to contractGovernor

This commit contains only the changes in the governance package. Changes to
Treasury and AMM are in separate commits, but tests won't pass without all
layers

Params in terms switch from arrays with embedded names to named records.
The electorate is passed to contracts in an invitation rather than a facet.

Dropped some tests in test-paramGovernance as the use of fakes was unsustainable.
   The two that were dropped were a happy path, and an inaccessible error path.

Added a test to show that changing the electorate works

* feat: add Electorate governance to the AMM

canonicalize the treatment of contract parameters

centralize initialization of AMM in tests.

* feat: add governable electorate to Treasury

* refactor: update install-on-chain

* chore: minor updates from review

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>