docs: clarify and correct writeup of XYK AMM README
Mostly this is clarification. I think this will make this easier for
me to read the next time I come back to it.

There was one correction: column 3 of the third table switched X and Y
in the 2nd and 3rd rows. You can tell that the update is correct
because the pool fee should be computed based on the unspecified
currency.