ATTACK HACK: allow vat code to arbitrarily corrupt messages
This enables user-level code to use E(target)._raw(methodname, corrupthook,
...args) instead of the usual E(target).methodname(...args), which will pass
'corrupthook' to m.serialize. In conjunction with similarly-evil changes to
marshal, this enables the creation of attack code that simulates a compromise
of liveslots, allowing the attacker to send arbitrary data into syscall.send.

Obviously this must never be committed to trunk. This is a test scaffold for
the Jan-2022 "purple team" attack exercise.