chore(swingset): improve typechecks
This enables typechecking for a few more swingset modules.

We've been a bit undecided about 'null' vs 'undefined'. Deliveries and
syscalls that cross the kernel/worker boundary (and get serialized into the
transcript, and unserialized again during replay) kind of want to use 'null',
because that's what JSON.parse emits. But of course optional parameters to a
function appear as 'undefined', and vatstoreGet('missing key') should return
'undefined'.

I'm going to aim for types that use 'null' in any sort of "optional" internal
context, to survive this, until/unless we define distinct types to represent
the values going over the wire or through transcript. But vatstoreGet will
continue to return undefined for missing keys.

I intend to add a test for vatstoreGetAfter later, and nail down the
null-vs-undefined question. I think syscalls should return things with
`undefined`, but the VatSyscallResult should be defined in terms of `null`.

To be robust, I did change one place in comms, where I was too picky and used
`s === undefined`, to use just `!s`.