chore(swingset): factor some test utilities out of util.js
test/util.js has a variety of utilities for unit tests, but was imported by
both start-compartment test programs, and vat code. The additional imports
from other packages cause a problem when bundling code for a vat.

This moves the handful of utilities we want for vat code into a separate
vat-util.js, which is then imported+rexported by util.js for the
start-compartment code.

It also moves some utilities from test-kernel.js into util.js for the benefit
of future unit tests.