fix(swingset): remove support for createVat(bundle)
This removes support for passing a full bundle to
`E(vatAdminService).createVat()`, leaving a bundlecap as the only remaining
option. Userspace is obligated to obtain a bundlecap as early as possible.

This can't land until Zoe/ertp/governance are changed to use bundlecaps, as
well as changing their unit tests to match.

Eventually I also want to remove `E(vatAdminService).createVatByName()`, and
have the code that needs it (zoe creating ZCF vats?) use
`getNamedBundleCap()` first.

closes #4486