test: update zoe+swingset-using tests to provide zcfBundle
This updates unit tests in (zoe, governance, run-protocol), specifically
those which use a swingset environment, to fetch the ZCF contract bundle from
the Zoe package, and install it into the kernel config they build. This is
necessary now that Zoe's `setupCreateZCFVat` no longer includes a copy.

Any new tests which follow this pattern will need to do the same.

refs #4487