feat: allow AMM liquidity provision at arbitrary rate
Add a new invitation that allows adding liiquidity in an arbitrary ratio.

The virtual pools get an internal facet that allows access to
addLiquidityActual.

In order to allow the rebalancing trade and liquidity addition to be
done in a single transaction, pool methods deal with stagedAllocations
rather than actual allocation.

closes: 4249