fix: when trades for zero are requested don't throw
I thought no one needed to request zero prices, and so I threw in that
case.  https://github.com/agoric/agoric-sdk/issues/4114 reveals a
situation where it will happen in ordinary circumstances.

Change the AMM code to respond with an empty amount when an empty
amount is offered, and vice versa.