fix: change all packages to call createVatAdminService the new way
The swingset fix for #4521 requires all callers of `createVatAdminService()`
to provide it with `devices.bundle` in addition to the existing
`devices.vatAdmin`. This occurs in the `bootstrap()` function, in both the
production configuration (`packages/vats/src/bootstrap.js`) and in
swingset-based unit tests in all packages.

This commit updates all agoric-sdk packages to do this.

refs #4521