fix(swingset): deliver startVat to setup() vats
For normal vats (both static and dynamic), `dispatch.startVat()` is the
signal to evaluate the userspace code bundle and invoke `buildRootObject`.
Liveslots has access to `syscall` earlier, but it must not use it until
`startVat()` indicates the kernel is ready (prepared to record the syscalls
in a transcript). However, `setup()`-based "raw vats" did not receive
`startVat`.

This changes the kernel to send `startVat` to `setup()`-based vats too. The
only real `setup()`-based vat we use is comms, which now ignores `startVat`,
but in the future will be able to use it to avoid constant vatstore
initialization queries.

Several unit tests create `setup()`-based vats: these were changed to ignore
`startVat` too. The test which counted message deliveries were updated to
accomodate the extra deliveries.

refs #4637
but doesn't close it until comms is improved too