fix(swingset): silently ignore attempts to kill a dead vat
Userspace can do `E(adminNode).terminateWithFailure()` multiple times on the
same vat. Previously, the kernel sent multiple `vatTerminated` messages to
vat-vat-admin, and since the first one removed vat-vat-admin's memory of the
dead vat, the second one made vat-vat-admin think that a static vat had died,
and it emitted a scary warning message.

The kernel was already catching duplicate termination attempts (and not
trying to clean up the already-deleted state), but the `notifyTermination()`
call was outside the check. This moves the call inside the check, so the
kernel won't emit the multiple calls.

It also changes vat-vat-admin's per-vat admin node to remember that
`terminateWithFailure` was called once already, and ignore subsequent calls.

closes #4866