fix(Swingset): add no-op dispatch.stopVat
This introduces `dispatch.stopVat`, a delivery made to vats that are
about to be upgraded. Liveslots ignores it for now, but soon it will
implement the cleanup phase of upgrade, where all non-durable data is
deleted and refcounts adjusted.

Comms ignores it too, but the comms vat does not use liveslots and we
can upgrade it without a cleanup step, since all state is already in
the database.

refs #1848