fix(swingset): upgrade provides vatID, calls stopVat
This changes processVatUpgrade() to perform a `stopVat` on the vat
about to be upgraded. It doesn't do any of the rest of the work yet,
so it always returns an error case.

It also fixes vat-vat-admin and friends to include the `vatID` in the
upgrade event, which was accidentally missing before.

refs #1848