chore(swingset): update vatstore-usage docs with new example IDs
This updates the object IDs used as examples to match the values they
ought to have for real, now that initialization is done early.

It also discusses "baggage" with more detail, and emphasizes the
surprising difference between the way virtual kinds and collections
use the second component of the vref.