fix(swingset): fix two doMoreGC-tracking bugs
Virtual collections `deleteInternal()` was ignoring the `doMoreGC`
result when dropping the value of a collection entry. This could fail
to re-run `gcAndFinalize` when necessary, allowing a Remotable (and
any in-memory objects it referenced) to survive longer than
appropriate before getting GCed.

Also, `deleteStoredVO()` allowed `doMoreGC` to short-circuit a call to
`removeReachableVref`, so if a Remotable appeared in the slots of a
virtual object's state, any slots after that point would not be
collected.

fixes #5044