fix(swingset): disable metering on both initial and from-snapshot workers
When vat options instruct swingset to disable metering in a vat, a bug
in `makeStartXSnap()` caused metering to be enabled anyways for the
second and subsequent worker used to host that vat. The first worker
starts from an empty XS engine, but after a few deliveries the
vat-manager will write out a heap snapshot, after which any new worker
will start from the snapshot. The `xsnap()` options that would disable
metering were not applied to the call that loaded a worker from
snapshot.

The consequence was that the first launch of a kernel would allow vats
to perform CPU-intense deliveries (as intended), but once the kernel
was restarted, those same delieries would terminate the vat (`compute
meter exceeded`) during transcript replay. If the snapshot was taken
recently enough that the remaining transcript did not include the
CPU-intense delivery, the new kernel would start, but the vat would
still be killed by any new delivery that exceeded the
unintentionally-imposed meter.

closes #5040