fix(swingset): only ceaseRecognition facets on virtual objects
`ceaseRecognition()` is called when an object is retired, to allow
weak-keyed collections to delete their corresponding entries. When a
cohort of facets is retired, it is called with a "baseRef" like
`o+145/3`, and then ceaseRecognition is responsible for enumerating
all the facets of that baseRef, so it can retire all the
pieces (e.g. `o+145/3:0` and `o+145/3:1`).

`ceaseRecognition` was not looking carefully enough at the properties
of the vref it was given, so when the kernel issued a
`dispatch.retireImports` on a normal object whose numeric ID happened
to collide with a locally-defined virtual Kind (e.g. `o-145`),
ceaseRecognition attempted to enumerate facets anyways. This caused it
to recurse with vrefs like `o-145:0` and `o-145:0:0`, etc. Eventually
this caused `parseVatSlot` to balk, making the `retireImports`
fail. This killed the vat.

The fix is to limit the facet-enumeration clause to vrefs which are
really virtual object baserefs: we must check the `allocatedByVat` and
`virtual` flags too.

fixes #5113