test: a virtual object test for AMM pools.
see: #5207

Pretty rudimentary.  It just sees that the objects come back. The test
fails when I manually remove (tested) stuff from state.

There's one test that changed state is persisted: I update the
notifier, and verify that on revival, the count has increased.

A plausible next test would be to add liquidity, but that's a bunch
more work.