fix(solo): introduce runStake attMaker to wallet admin facet
fixes #4263

This addresses a gap in the runStake user flows where a
user would have to use the REPL to do:

 E(E(home.wallet).getAdminFacet()).resolveAttMaker(home.attMaker)

We use setPresences to detect the arrival of attMaker on the chain
bundle (decentralObjects) and setWallet to detect the arrival of the
wallet. When both are available, we introduce them with
resolveAttMaker.