feat: make amm.addPool() require minimum collateral
When creating a new pool for the AMM, there will be a minimum amount
of IST. The caller is expected to provide an equivalent value of the
collateral to balance the AMM pool. The liquidity tokens corresponding
to that minimum will be added to the AssetReserve. If the caller adds
more liquidity than the minimum, they get the extra liquidity tokens.

Closes: #4643

This required updating a number of tests to initialize their AMMs with
a minimal amount.