feat(AMM)!: make amm.addPool() require minimum collateral (#5377)
* feat: make amm.addPool() require minimum collateral

When creating a new pool for the AMM, there will be a minimum amount
of IST. The caller is expected to provide an equivalent value of the
collateral to balance the AMM pool. The liquidity tokens corresponding
to that minimum will be added to the AssetReserve. If the caller adds
more liquidity than the minimum, they get the extra liquidity tokens.

Closes: #4643

This required updating a number of tests to initialize their AMMs with
a minimal amount.

* fix: don't create AMM pools in `addAssetToVault`

* docs: update references to addPool in documentation

* fix: update demoIssuers to use addPoolInvitation.

* refactor: turn minInitialPoolLiquidity into an Amount

* chore: cleanups.  variable names and typescript annotations.

* fix pool helper types

* parameterize VirtualPool type

* pool amounts are nat

* chore: update TODO to include issue #

* test: uncomment a test assertion before pool creation

* refactor: integrate with #5397

update interchainPool tests

switch minInitialPoolLiquidity back to Nat in startupAmm since it has
to be driven from text values in environment variables.

* chore: lint errors (unreferenced names)

Co-authored-by: Turadg Aleahmad <turadg@agoric.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>