fix(swingset): fix handling of promises in virtual data
Previously, liveslots managed the lifecycle of VPID (promise ID)
registration as if the only consumer was the kernel: every time we
serialized a promise, we assumed it was being shared with the kernel,
so we would use `.then` to prepare a `syscall.resolve` message, and
we'd retire the VPID<->`Promise` mapping immediately upon resolution.

But #5106 reveals that storing a Promise in virtual data introduces a
different lifecycle. The VPID/Promise registration needs to survive as
long as 1: any virtual data references the promise, or 2: the kernel
is aware of the VPID. The latter happens either through "import" (the
kernel is the decider, and the vat's `Promise` is a follower, waiting
for `dispatch.notify`), or "export" (the vat is the decider, waiting
to do a `syscall.resolve`).

This changes the state machine for VPIDs, to maintain the registration
as long as necessary. VPIDs are tracked in two Maps named
`exportedVPIDs` and `importedVPIDs`.

A new virtual table (`makeScalarBigMapStore`) is used for Promises
currently being tracked by `watchPromise`, to keep the registration
alive until they are resolved or the vat is upgraded. This causes some
test churn as the identifiers for other tables are now one larger.

fixes #5106