fix(swingset): use dynamic vat options.name for the xsnap no-op argument
Previously, static vats would glean a "name" value from the config
structure, and dynamic vats would not get a name. Both kinds of vats
would glean a "description" that combined this name, a `description`
argument provided to `createVat()`, and information about the source
bundle used to create the vat. If `name` was available, it would be
used as an extra no-op argument to `xsnap`, so by looking at `ps`, you
could tell which `xsnap` worker process was hosting which vat.

As we move to bundlecaps (especially as the chain bootstrap code moves
to bundlecaps), we've lost the `name` field in `ps` for most initial
vats.

This change allows the creator of a dynamic vat to provide a `name`
field, just like the config structure can provide then for static
vats. This `name` is now uniformly used in the `ps` arguments.

The `description` option has been removed. Slogfiles still include a
description, and it incorporates the assigned name and source bundle
information, but userspace can no longer provide
`E(vatAdminSvc).createVat(bcap, { description })` in the options.

The `name` field is arbitrarily limited to 200 characters, to avoid
running into OS limitations when `xsnap` includes in in `argv`. We
also limit it mostly identifier-ish characters, to avoid confusion
with spaces and control characters, etc.

fixes #5516