fix(xsnap): fix test which reused a mutable ArrayBuffer
test-replay.js was providing a mock `writeFileSync()` to record the
replay records in RAM instead of on disk. However the caller is using
the same ArrayBuffer each time, and the mock was holding onto the
buffer, not the contents, which means it wound up with the last thing
written (corrupted by subsequent writes). The `transcript1` comparison
data was corrupted in a matching way.

This fixes the mock to make a copy of the buffer, and updates the
comparison data to be the real expected string ("Hello, World!").