fix: minor tweak to make `sortObjectProperties` use `hasOwn`
Yesterday's PR #5552 originally used `Object.hasOwn` but the XS engine lacked
the `Object.hasOwn` function and so all the XS-based tests failed, so the code
was tweaked to do something lesser.  But today a new version of XS appeared that
does have `hasOwn` so this small PR puts things to right.