fix: eliminate `fakeDurable` and change global durability rules to match
This addresses the SwingSet portion of #5593.

-- Virtual objects and collections can no longer be declared `fakeDurable`; they
   are either durable or not, as was the case prior to the introduction of the
   `fakeDurable` feature.

-- The `enableFakeDurable` config flag is renamed to `relaxDurabilityRules`.

-- This renamed flag defaults to `false` when running in a swingset (as before)
   and to `true` when running on top of the mock virtuals library provided by
   `fakeVirtualSupport.js` (in the latter case it previously was unconditionally
   `true`; now it is settable as an option to `makeFakeVirtualStuff` et al).

-- The meaning of the flag is altered.  By setting it to `true`, rather than
   allowing `fakeDurable` to be used or not, it instead globally alters the
   rules of what values can be admitted to durable objects and collections to be
   the same as their non-durable counterparts.

The expectation is that production code and tests that are exercising vat
upgrade scenarios will run with the flag set to `false`, whereas tests that are
exercising business logic that do not care about durability and which use mocks
that won't pass the durability checks anyway will run with the flag set to
`true`.  This expectation is reflected in the default flag values adopted by the
kernel (`false`) and the fake virtual support code (`true`) respectively.