fix(swingset): don't double-allocate durable object vrefs
The `nextInstanceID` for durable objects, used to allocate vrefs for
new instances, was held ephemerally in RAM. As a result, the first
call to `makeFoo()` (for `makeFoo = defineDurableKind(fooHandle..)`)
in version-2 would get the same vref as the first `makeFoo()` from
version-1, when these should be entirely distinct objects. In
practice, the new object tended to displace the old one.

This changes the virtual object manager to track `nextInstanceID`
durably, by writing it into the `vom.dkind.${kindID}` metadata. This
metadata now holds `{ kindID, tag, nextInstanceID }`. We write this
out after every durable object allocation (previously, the metadata
was only written once, during `makeKindHandle`). Non-durable (merely
virtual) objects do not include this counter in the metadata, nor do
they write it to DB more frequently than before.

The upgrade test was modified to look for this overlapping vref
problem.

closes #5725