fix: provide policyInput=none for non-delivering cranks
Originally, we set `policyInput = ['none']` when a run-queue event did
not result in a vat delivery (e.g. sends to a terminated vat, or
promise-resolution notifications that were previously delivered as
part of a resolution-graph batch). This branch regressed on that
change slightly, using `['crank']` (with no computrons) even when no
delivery was made.

This commit fixes the branch to behave more like the original:
`['none']` when no delivery is made. It also adds computrons to
`['vat-terminated']`, which previously lacked computrons: this should
give the run policy more accurate information to make runtime
estimates.