fix: serialize all tests in test-terminate.js
The 32 tests in SwingSet's test-terminate.js each launch 5-6 xsnap
workers, and if we let them all run in parallel, the resulting spike
in memory usage was triggering the OOM Killer. Local testing shows an
8GB host cannot run everything in parallel, but a 16GB host can. We
think GitHub's CI hosts are closer to the lower end, because we've
seen CI jobs fail with a mysterious "test-terminste.js killed by
SIGKILL" message in the logs.

This commit forces all tests to run in serial. On my home machine it
slows test-terminate.js down from 48s to 68s, but reduces the memory
spike by a lot.

closes #5782