docs: a set of diagrams on Zoe and Zcf datastores and initialization
refs: #4383

The datastore diagram was created in Miro:
https://miro.com/app/board/uXjVPR6j4Nc=/?share_link_id=559329933746

The others are .puml rendered to png.

These diagrams show the state of Zoe and ZCF as of 2022/10/1. I expect
to update them as I finish the conversion to durability.

The datastore diagram shows which objects and stores are accessible to
Zoe or to ZCF, and which are shared.

The sequence diagrams show the order in which the various objects are
created, passed between the vats, and when promises are resolved.