chore(deps): update `source-map@^0.7.3` to 0.7.4
The `source-map@0.7.4` update is important for enabling Node.js v18
compatibility. Older versions use a `typeof fetch === 'function'`
check to enable browser-specific behavior. With this release however,
the browser check is moved to whether `window` is defined and matches
`this` value at check time. Node.js 18 has a `fetch` function, so
older `source-map` versions incorrectly flag it as 'browser'. None
of Node.js versions have a `this === window` situation going on.

This is a change on the path to Node.js 18 compatibility route,
in anticipation of #6489 offering official support down the line.