chore: make krefOf/extractSingleSlot more picky
When krefOf() is called as part of kmarshal.serialize, marshal will
only give it things that are 'remotable' (Promises and the Far objects
created by kslot()).  When krefOf() is called by kernel code (as part
of extractSingleSlot() or the vat-comms equivalent), it ought to throw
if 'obj' is not one of the Far objects created by our kslot().

This also changes extractSingleSlot() to be just as precise as the old
implementation, to be safe against future changes to krefOf() or the
marshalling format (e.g. #2069 auxdata adding additional properties).