fix(swing-store): move getAllState/setAllState onto a debug object
`getAllState` and `setAllState` are swing-store helper functions which
copy (or set) all the state of a store at once, used exclusively for
testing.

Previously, they were exported by swing-store, and applied to a
`kernelStorage` object, but that depended upon a
`streamStore.dumpStreams` method that we'd rather not expose to normal
users.

This commit adds a new `debug` object, next to `kernelStorage` and
`hostStorage`, which holds `getAllState`, `setAllState`, and
`dumpStreams`. The `kernelStorage.streamStore` object no longer has a
`dumpStreams` method.

This also updates all the swingset unit tests which were using the
helper functions.