fix(swing-store): replace kvStore.getKeys() with getNextKey()
The kernel uses some portions of the kvStore as a list or map, and
occasionally needs to iterate through that collection, for example
when deleting all vatstore entries for a terminated vat, or examining
all c-list entries during upgrade to find promises for
cancellation. This feature is shared with vats as well (through the
`vatstoreGetAfter` syscall), so userspace can iterate through virtual
collections with .keys() and .values().

Previously, swing-store enabled this through a method named
kvStore.getKeys(), which accepted a starting point (inclusive) and an
optional ending point (exclusive). This mapped closely to the
iteration API provided by the original LMDB backend. When we moved to
SQLite, which offers plenty of queries but not that kind of iterator,
swing-store implemented getKeys() by storing a full copy of the
requested range in RAM, and then yielding values from it one at a
time. This induces a (potentially large) memory burden, for however
long the kernel chooses hold on to the iterator. Under some
circumstances, userspace can provoke this burden too.

This commit changes the swing-store API to provide a getNextKey()
instead of getKeys(). The getNextKey() API takes `previousKey`, which
may or may not exist, and returns the next (lexicographically ordered)
key, which may or may not exist. Kernel and userspace can build an
iterator on top of this as needed. Each call to `getNextKey()`
corresponds to exactly one DB query, and no state is retained between
queries.

Possible routes for optimization/improvement:

* getNextEntry() -> (nextKey, nextValue)
  * Many of our use cases are for deletion, and don't need the value,
    so I didn't want to start here, but it would reduce the number of
    syscalls and queries for the cases that do need the value.
* getNextKey(priorKey, maxKey)
  * This would return `undefined` if the next key were greater than
    `maxKey` (perhaps with variations for inclusive vs exclusive).
    I don't think this would reduce the number of calls, but it might
    make some of the syscall logs more legible (kernel/userspace would
    received a fixed `undefined` on their last query, instead of an
    unrelated value which they would then filter out).

refs #6468