feat(swingset): rewrite kernel to use getNextKey instead of getKeys
This rewrites the kernel to use the new swing-store `getNextKey()`
API, instead of the old `getKeys()` iterator. It also changes
`syscall.vatstoreGetAfter()` to `vatstoreGetNextKey()`, which maps
directly to the swingstore API (plus some range enforcement).

As a result, we can remove the awkward adapters and kernelSyscall.js
code which kept track of which vat was holding onto the one swingstore
iterator (which needed to clear that iterator if/when some other vat
did a vatstoreGetAfter).

Liveslots acquired `vatstore-iterators.js` with helper functions, so
the virtual-collection code can continue to iterate through ranges of
keys, on top of `vatstoreGetNextKey`.

Some virtual object/collection tests were overly sensitive to the
exact syscalls being provoked, and this patch caused a lot of churn,
so I stripped down those tests to only examine the resulting vatstore
state that we really care about. This deleted a lot of test code.

closes #6468