feat: move snapstore into SQLite database with the rest of the swingstore
This is phase 1 of #6742.  These changes cease storing snapshots in
files but instead keep them in a new table in the swingstore SQLite
database.  However, in this commit, snapshot tracking metadata is
still managed the old way using entries in the kvstore, rather than
being integrated directly into the snapshots table.