fix: make test-activityhash.. use xs-worker
I'm seeing CI failures under Node-18 in this test, presumeably because
we get GC variation between the first run and the replay. Super
annoying, and not germane to what this test is supposed to be
executing. The `test.serial` wrapper wasn't enough to fix it, so I've
just switched the test to only use xs-worker.

refs #3240