fix(swingset): remove unused worker types: node-subprocess, nodeworker
Most of our vat workers use `xs-worker`, which runs the code in an
`xsnap` subprocess. A handful of vats (maybe just comms) run in a
`local` worker, which is co-resident with the kernel.

We have two additional worker types, `nodeWorker` (which uses threads,
known as "workers" in Node.js), and `node-subprocess` (which runs a
second copy of Node.js in a subprocess). I added these for
completeness, and to keep us honest about handling different worker
types. But they are incomplete (node-subprocess *could* block for a
syscall, but I didn't implement that),
underfeatured (nodeWorker *cannot* block for a syscall without heroics
involving ShareArrayBuffers and Atomics), undertested (the previous
lacking features prompted us to `test.skip` them), and unused by
anything outside of our test suite.

I have refactoring work in the pipeline that will be easier if there
is less code to move around. So it's time for these to go. I might try
to bring these back some day, but only when we have some interesting
use case for them (which is more likely to be inside a browser, than
under Node.js).

closes #4884