fix: bootstrap handles BundleIDs, not full bundles
The core-bootstrap `installBootContracts()` step is responsible for
creating Zoe installations for each "boot contract": contracts that
are included in the initial chain configuration, and instantiated
during bootstrap (currently just `centralSupply` and
`mintHolder`). These contract bundles are specified by config.bundle
entries, which provide a `sourceSpec` that points to the bundle file
on disk.

Previously, this step obtained a full bundle for each contract, in
RAM, and then used `zoe.install(bundle)` to create the Zoe
Installation. When called this way, Zoe must store the full
bundle (about 1MB) in its durable storage, and the bundle appears as a
message argument in lots of places on its way to being evaluated
inside a ZCF vat, which slows things down and consumes more disk
space.

The preferred way to interact with bundles is through BundleCaps. This
commit changes installBootContracts to look up the BundleID of each
boot contract, and submit the ID to `zoe.installBundleID()`. This
allows Zoe to retrieve a (small) BundleCap, and store the bcap in the
installation record. With this approach, the full bundle only appears
once, when the ZCF vat retrieves its contents just prior to
evaluation.

Changing installBootContracts is straightforward, however we have a
number of mock Zoe and mock VatAdminService implementations that
appear during unit tests, and these required changes to accomodate the
new approach. `zoe/tools/fakeVatAdmin.js` provides the primary mock
service, but it is only capable of creating ZCF vats. It was augmented
to understand the full suite of `getBundleIDByName`/etc methods, with
helper tools to populate the simulated tables. Then
`vats/tools/boot-test-utils.js` acquired a wrapper around zoe's mock
service, adding the ability to create non-ZCF vats, and populating the
tables with all the bundles known to the unit tests.

refs #6826
refs #4374