fix: replace zoe.install with zoe.installBundleID
Many governance proposals are deployed as bundles of code to evaluate
on-chain in the core-bootstrap environment, known as executing the
"Big JS Hammer". This might reference contracts (which must
necessarily be installed in Zoe first), but the behavior code itself
is neither a contract nor a vat.

Previously, core-bootstrap (ab)used Zoe to hold and name these bundles
of behavior code: they were sent to `zoe.install(bundle)`, the
installation handle was put on the Board, and the Board name was used
in the execution command. One consequence is that large code bundles
were passed in multiple messages, including a lot of vatstore syscalls
as Zoe saved the full bundles in virtual objects.

This commit changes the pathway to use BundleIDs (hashes) and
BundleCap objects, rather than full bundles, to reduce the cost and
overhead of sending large code bundles around. And the behavior code
is removed from Zoe entirely, leaving Zoe solely in charge of
contracts.

chain-behaviors.js was updated to produce `evaluateBundleCap`, rather
than `evaluateInstallation`. In deploy-script-support/,
coreProposalBehavior was changed to make `restoreRef` use either
bundle names (as installed in config.bundles, in test / simulation
environments, where the deployment is known at the time of chain
launch, or for proposals that are included in the launch of the real
chain), or BundleIDs (for when the deployment is installed to the
chain with transactions, after launch).

In all places where the code was using `bundleID` to refer to a
generated name for the bundle (e.g. as keys of `config.bundles`), it
was changed to use `bundleName`. We reserve the term "bundle ID" for
the SHA256 hash-based identifier of a bundle, as is the practice
throughout the kernel code and the vat-admin API.

The writeCoreProposal.js code which is responsible for publishing bundles
during the deployment-creation phase has been temporarily disabled, as
it needs to be rewritten to generate BundleIDs instead of full
bundles.

Several tests which depended upon behavior code being installed in zoe
were changed to match the new scheme. The fakeVatAdmin.js tool was
enhanced to provide a copy of `vatPowers` with a `D()` that can handle
`D(bundleCap).getBundle()`, necessary for those tests.

refs #6826