feat: create new xsnap-lockdown package
This extracts the "lockdown" bundle out of packages/xsnap and into a
new package whose NPM name is `@agoric/xsnap-lockdown`, and lives (for
now) in packages/xsnap-lockdown .

The bundle is generated explicitly, by running `yarn build`. The
intention is that each published version of `@agoric/xsnap-lockdown`
has a stable bundle, with specific (hashable) contents. Importing
clients can rely upon this stability to e.g. have deterministic xsnap
heap snapshot contents even if the versions of other parts of the
system have changed.

The new README.md describes the API, the stability goals, and how we
achieve them.

Fairly soon, we'll need manage the version number of this package more
explicitly, either by moving it to a different git repository, or
changing out use of `lerna publish` to avoid spurious changes. But for
now we can leave it in place.

refs #6596