feat: extract swingset-xsnap-supervisor out to a separate package
This extracts the XS supervisor bundle out of packages/SwingSet and
into a new package whose NPM name is
`@agoric/swingset-xsnap-supervisor`, and lives (for now) in
packages/swingset-xsnap-supervisor .

The bundle is created explicitly, by running `yarn build`. As with
`xsnap-lockdown`, the intention is that each published version has a
stable bundle, with specific (hashable) contents. Importing clients
can rely upon this stability to e.g. have deterministic xsnap heap
snapshot contents even if the versions of other parts of the system
have changed.

The new README.md describes the API, the stability goals, and how we
achieve them.

Fairly soon, we'll need manage the version number of this package more
explicitly, either by moving it to a different git repository, or
changing our use of `lerna publish` to avoid spurious changes. But for
now we can leave it in place.

refs #6596