fix(swingset): move a bunch of types from swingset to swingset-liveslots
These types define the messages (deliveries and syscalls) that can
pass between a swingset kernel and one of its workers. By moving them
to swingset-liveslots, they can be imported by the upcoming
"swingset-xsnap-supervisor" package.

Strictly speaking, these are not defined by (or limited to) liveslots:
we can imagine alternate (non-liveslots) vat runtimes which still
speak the same language of deliveries and syscalls. For example we
might define a less object-capability -ish runtime whose userspace
code makes direct vatstore changes, in more of an ORM style, which
lacks internal confinement but might be faster/cheaper. Or we could
use a language other than JavaScript.

In the long run, we probably need to define a separate
"swingset-vat-api-v1" package, which defines the kinds of deliveries
and syscalls that can be exchanged between the kernel and a vat
worker. The kernel-side translators will emit these, the worker
management code will transport them, and they will be consumed by
liveslots/etc within the worker.

If we add a new kind of delivery in the future, we could define
"swingset-vat-api-v2" with the additions, the kernel could use both
-v1 and -v2, and the translator would refrain from sending the extra
-v2 messages to an older worker.