fix(swing-store): remove explicit wal_checkpoint
We're using `synchronous=FULL`, so we don't need to perform an
explicit `wal_checkpoint(FULL)` to obtain durability in the face of
power loss. And the FULL-mode checkpoints cause a 5s busy-wait when we
have an external reader connection (like the upcoming swing-store
exporter, which runs in a thread or subprocess, with it's own DB
connection).

This commit just removes the call to `PRAGMA wal_checkpoint(FULL)`
entirely.

closes #7069