feat(auction): add an auctioneer to manage vault liquidation (#7000)
* feat(auction): add an auctioneer to manage vault liquiditation

Separate pieces include a scheduler that manages the phases of an
auction, the AuctionBook that holds onto offers to buy, and the
auctioneer, which accepts good for sale and has the APIs. Params are
managed via governance. The classes are not durable.

This is DRAFT, for early review.

The changes to VaultFactory to make use of this approach to
liquidation rather than the AMM are in a separate PR, which will be
available before this PR is finalized.

closes: #6992

* refactor: cleanups from working on vault liquidation

* chore: cleanups suggested in review

* refactor: simplify keys in sortOffers

* chore: cosmetic cleanups from review

* chore: auction was introduced to vaultfactory too early.

It will be introduced with #7047 when vaultfactory makes use of the
auction for liquidation

* chore: correctly revert removal of param declaration

It'll be removed in #7074, I think, but for now it like a change in
this PR.

* test: repair dependencies for startAuction

* chore: responses to reviews

* test:  refactor proportional distribution test to use macros

* refactor: correct sorting of orders by time

Add a test for this case.

make all auctionContract tests serial

* chore: cleanups suggested in review

for loops rather than foreach
better type
extraneous line

* chore: clean up scheduler; add test for computeRoundTiming

* chore: minor cleanups from review

* chore: rename auctionKit to auctioneerKit

* chore: drop auction from startPSM

* chore: clean-ups from review

* refactor: computeRoundTiming should not allow duration === frequency

* chore: clean up governance, add invitation patterns in auctioneer

* refactor: don't reschedule next if price is already locked