fix: add ensureTxn to (most) snapStore operations
We need this in every method that might perform a SQL write, to
protect the long-lived SQL transaction that is managed by the host
application (via `hostStorage.commit`). The kvStore and
transcriptStore operations had them, but they were missing from
snapStore.

We omit ensureTxn() from the debug.dump() methods (dumpKVEntries,
dumpTranscripts, dumpSnapshots), because these are typically run from
unit tests, outside the normal txn, and we don't want to leave the DB
in a locked state after they finish.