swingstore: importer should call exporter.close()
We defined close() on the Exporter type, and implemented it on the
exporter, and we need it to be called: it closes the DB connection and
discards any (accidental) changes that the (read-only) exporter might
have made.

However our importer was not calling close() on the exporter it
receives.

This changes the importer to call close(). It also changes a unit test
to not close the exporter on the importer's behalf.

Note that this means exporters are consumed by the importSwingStore()
process.