liveslots: add a new cache utility
This cache accumulates all data until an explicit flush, at which
point it writes back all dirty state, and forgets all state.

This avoids the somewhat unpredictable behavior of an LRU cache, which
caused us problems with GC sensitivity. It suffers from the potential
for larger RAM usage between flushes, but is quite simple, and yields
easy-to-debug syscall behavior.

We have other caches in our platform, but liveslots needs something
with these specific properties, and I don't know what the needs in the
rest of the platform are, so I'm fine with not sharing this
implementation with other packages, at least for now.

refs #6693