enforce that buildRootObject() returns an ephemeral, not durable
We assign the special "o+0" vref to the root object. Newly created
virtual/durable objects are assigned a vref like "o+d13/5", which
incorporates the KindID and the instance sequence number.

If buildRootObject() creates a virtual/durable object and returns it
as the root, we now have one object with two different identities,
violating the main slotToVal/valToSlot invariant. The initial
Representative will have a context and a state that was created and
cached under the durable vref, but by the time a method is called,
valToSlot knows it as "o+0", so the vatstoreGet state lookup fails (as
of course there is no state stored under "o+0").

This commit inserts a liveslots check to ensure that the root object
does not already have a vref.