feat(swingset): persist xsnap bundle IDs at vat creation/upgrade
When launching an xsnap worker, we must load two special bundles into
the worker's empty JS engine: xsnap-lockdown (which turns it into
a SES environment), and swingset-xsnap-supervisor (which loads
liveslots and the command handlers). We get these bundles from
packages of the same names.

Previously, we would load the current versions of these bundles each
time the controller/kernel started, and use them for all vats started
or upgraded during that incarnation of the controller (i.e. until the
process exited). This did not maintain stable behavior for a vat which
e.g. does not record a heap snapshot before its initial controller
incarnation finishes: if the xsnap-lockdown package were updated
between controller incarnations, the second incarnation would launch
the worker with a different bundle than the first, potentially causing
divergence between the vat's original behavior (as recorded in its
transcript) and the replay. If there were any local-manager liveslots
-based vats, these would suffer the same problem on every kernel
reboot, since only xsnap-hosted vats can use heap snapshots.

With this commit, the lockdown and supervisor bundles are stored in
the swing-store `bundleStore`, and their IDs are recorded in the
metadata for each vat. We check the bundle-providing packages each
time we start a new vat (or upgrade an existing one), so the new
worker will use the latest bundles. But by recording the IDs
separately for each vat, we ensure that every launch of that vat
worker will the same bundles. This stability will help us maintain the
deterministic behavior of a vat despite changes to the packages
depended upon by any given version of the kernel.

Unit tests and replay tools can use `options.overrideBundles=[..]` to
replace the usual bundles with alternates: either no bundles, or
custom lockdown/supervisor bundles.

closes #7208