add 'xsnap' as an externally-selectable managerType
External callers indicate their choice of vat worker with a "worker
type" or "manager type" string, through APIs like:
* `$SWINGSET_WORKER_TYPE` (aka `env.SWINGSET_WORKER_TYPE`)
* `config.defaultManagerType`
* `config.vats[NAME].creationOptions.managerType`
* the `options.managerType` passed to `E(vatAdminSvc).createVat()`

Previously, the external options were `local` or `xs-worker`. I'd like
to move us to using `xsnap` everywhere (it's shorter and more
accurate). The previous commits on this branch change the internal
usage (the per-vat `options` metadata, and `workerOptions.type`) to
use `local` or `xsnap`. But the external options remained the same.

This commit adds `xsnap` as an option to the external API, doing the
same thing as `xs-worker`. This lets us claim that `xsnap` is the
"right" spelling, and `xs-worker` is an alias.

Later, we can change a lot of callers (hopefully all) to use `xsnap`,
and add a deprecation warning to catch the stragglers, and eventually
remove the alias.