remove in-RAM list of all collections
This removes `deleteAllVirtualCollections`, and the in-RAM
`allCollectionObjIDs` Set of all collectionIDs which supported
it. Once upon a time, `stopVat` called `deleteAllVirtualCollections`
to delete the vatstore data for all non-durable virtual
collections. We removed responsibility for deleting virtual data from
the vat (and stopped calling `stopVat` altogether) in the interests of
reducing upgrade risk, and improving upgrade performance, despite the
fact that this will cause a DB storage leak during upgrade (we figure
we can find a way to fix that later, and disk is cheap).

Now that `deleteAllVirtualCollections` is gone, we don't need to spend
the RAM on each collection, which violated our high-cardinality design
rules anyways.

refs #5058