test(vats): use test.after.always to shutdown kernel
Several swingset-using tests in packages/vats use a `t.after` hook to
call `controller.shutdown()` at the end of the test. This is important
to do, because the kernel/controller will have child `xsnap` processes
running until it is shut down, and V8 won't exit until they're
gone. This would cause a timeout error.

It turns out that `t.after` hooks do not run if any of the tests have
failed. So if any of the packages/vats/ tests were to fail, we'd get a
timeout error instead of the proper error message, and the error would
be delayed by some arbitrary amount (currently about 30 seconds, via
the patch Mathieu applied to AVA).

This changes our `t.after(` calls to use `t.after.always()` instead,
which *does* run even after tests have failed. We're hoping this might
fix some of the hanging tests we've been seeing, but even if it
doesn't, `t.after.always` is clearly the right way to do it.